refactor: HTTP 메세지 관련 클래스 패키지 이동
