+++ src/main/java/annotations/HandleRequest.java
@@ -1,6 +1,6 @@
 package annotations;
 
-import webserver.HttpMethod;
+import webserver.httpmessage.HttpMethod;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
