+++ src/main/java/webserver/APIHandler.java
@@ -1,6 +1,9 @@
 package webserver;
 
 import annotations.HandleRequest;
+import webserver.httpmessage.HttpMethod;
+import webserver.httpmessage.HttpRequest;
+import webserver.httpmessage.HttpResponse;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
