+++ src/main/java/webserver/RequestHandler.java
@@ -5,6 +5,8 @@ import java.net.Socket;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.httpmessage.HttpRequest;
+import webserver.httpmessage.HttpResponse;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
