+++ src/main/java/webserver/httpmessage/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.httpmessage;
 
 import exceptions.HttpRequestFormatException;
