+++ src/main/java/webserver/httpmessage/HttpRequest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.httpmessage;
 
 import utils.IOUtils;
