+++ src/test/java/webserver/HttpRequestTest.java
@@ -1,13 +1,9 @@
 package webserver;
 
-import exceptions.HttpRequestFormatException;
-import webserver.HttpRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
-
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
