+++ src/test/java/webserver/URITest.java
@@ -2,11 +2,11 @@ package webserver;
 
 import exceptions.HttpRequestFormatException;
 import org.junit.jupiter.api.Test;
+import webserver.httpmessage.URI;
 
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class URITest {
     @Test
